+++ lib/helpers.ts
@@ -2,6 +2,22 @@
 
 import type {Rule} from 'eslint';
 
+const topLevelTypes = new Set([
+  'ArrayExpression',
+  'AssignmentExpression',
+  'BinaryExpression',
+  'BlockStatement',
+  'CallExpression',
+  'ExportNamedDeclaration',
+  'ExpressionStatement',
+  'LogicalExpression',
+  'NewExpression',
+  'ObjectExpression',
+  'UnaryExpression',
+  'VariableDeclaration',
+  'VariableDeclarator'
+]);
+
 export function IsCommonJs(node: Rule.Node) {
   while (node.type !== 'Program') {
     node = node.parent;
@@ -11,7 +27,7 @@ export function IsCommonJs(node: Rule.Node) {
 
 export function isTopLevel(node: Rule.Node) {
   let scope = node.parent;
-  while (scope.type === 'BlockStatement') {
+  while (topLevelTypes.has(scope.type)) {
     scope = scope.parent;
   }
   return scope.type === 'Program';